+++ db_stress_tool/db_stress_shared_state.cc
@@ -1,0 +1,19 @@
+//  Copyright (c) 2011-present, Facebook, Inc.  All rights reserved.
+//  This source code is licensed under both the GPLv2 (found in the
+//  COPYING file in the root directory) and Apache 2.0 License
+//  (found in the LICENSE.Apache file in the root directory).
+//
+// Copyright (c) 2011 The LevelDB Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file. See the AUTHORS file for names of contributors.
+//
+
+#ifdef GFLAGS
+#include "db_stress_shared_state.h"
+
+namespace rocksdb {
+const uint32_t SharedState::UNKNOWN_SENTINEL = 0xfffffffe;
+// indicates a key should definitely be deleted
+const uint32_t SharedState::DELETION_SENTINEL = 0xffffffff;
+}  // namespace rocksdb
+#endif  // GFLAGS
